Set time to expire stale locks

+++ src/nexusformat/nexus/lock.py
@@ -34,7 +34,7 @@ class NXLock(object):
         File descriptor of the opened lock file.
     """
 
-    def __init__(self, filename, timeout=60, check_interval=1):
+    def __init__(self, filename, timeout=60, check_interval=1, expiry=28800):
         """Create a lock to prevent file access.
 
         This creates a lock, which can be later acquired and released. It
@@ -51,11 +51,15 @@ class NXLock(object):
         check_interval : int, optional
             Number of seconds between attempts to acquire the lock,
             by default 1.
+        expiry : int, optional
+            Number of seconds after which a lock expires, by default 8*3600.
+            Set to 0 or None to make the locks persist indefinitely.
         """
         self.filename = os.path.realpath(filename)
         self.lock_file = self.filename+'.lock'
         self.timeout = timeout
         self.check_interval = check_interval
+        self.expiry = expiry
         self.pid = os.getpid()
         self.fd = None
 
@@ -64,7 +68,7 @@ class NXLock(object):
                 + os.path.basename(self.filename)
                 + "', pid=" + str(self.pid)+")")
 
-    def acquire(self, timeout=None, check_interval=None):
+    def acquire(self, timeout=None, check_interval=None, expiry=None):
         """Acquire the lock.
 
         Parameters
@@ -75,6 +79,9 @@ class NXLock(object):
         check_interval : int, optional
             Number of seconds between attempts to acquire the lock,
             by default `self.check_interval`.
+        expiry : int, optional
+            Number of seconds after which a lock expires, by default
+            `self.expiry`.
 
         Raises
         ------
@@ -89,6 +96,9 @@ class NXLock(object):
 
         if check_interval is None:
             check_interval = self.check_interval
+
+        if expiry is None:
+            expiry = self.expiry
 
         timeoutend = timeit.default_timer() + timeout
         initial_attempt = True
@@ -105,13 +115,14 @@ class NXLock(object):
                 if e.errno != errno.EEXIST:
                     raise
                 # Remove the lockfile if it is older than one day
-                elif initial_attempt:
-                    if self.is_stale():
+                elif initial_attempt and expiry:
+                    if self.is_stale(expiry=expiry):
                         os.remove(self.lock_file)
                     initial_attempt = False
                 time.sleep(check_interval)
         # Raise an error if we had to wait for too long
         else:
+            self.fd = None
             raise NXLockException(
                 f"'{self.filename}' is currently locked by an external process"
                 )
@@ -190,9 +201,18 @@ class NXLock(object):
                                       "by an external process")
         return
 
-    def is_stale(self):
-        """Return True if the lock file is older than one day."""
-        return ((time.time() - os.path.getmtime(self.lock_file)) > 86400)
+    def is_stale(self, expiry=None):
+        """Return True if the lock file is older than one day.
+
+        If the lock file has been cleared before this check, the
+        function returns False to enable another attempt to acquire it.
+        """
+        if expiry is None:
+            expiry = self.expiry
+        try:
+            return ((time.time() - os.path.getmtime(self.lock_file)) > expiry)
+        except FileNotFoundError:
+            return False
 
     def __enter__(self):
         return self.acquire()
